+++ libs/dot-components/src/lib/components/css-grid/CssGrid.spec.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render } from '../../testing-utils';
+
+import { CssGrid } from './CssGrid';
+
+describe('CssGrid', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<CssGrid />);
+    expect(baseElement).toBeTruthy();
+  });
+});