+++ libs/dot-components/src/lib/components/index.ts
@@ -47,3 +47,6 @@ export * from './table/TablePagination';
 export * from './tabs/Tabs';
 export * from './button-toggle/ButtonToggle';
 export * from './typography/Typography';
+export * from './css-grid/CssGrid';
+export * from './css-grid/CssCell';
+export * from './css-grid/CssGridDebug';